feat: move heroBannerTitle component  to folder and add test
