+++ src/components/molecules/DetailedMovieCard.tsx
@@ -1,7 +1,7 @@
 import { FC } from "react";
 import { Card, CardMedia, CardContent } from "@mui/material";
 import { Movie } from "../../types/Movie";
-import HeroBannerTitle from "../atoms/HeroBannerTitle";
+import HeroBannerTitle from "../atoms/HeroBannerTitle/HeroBannerTitle";
 import HeroBannerDescription from "../atoms/HeroBannerDescription/HeroBannerDescription";
 import HeroBannerRating from "../atoms/HeroBannerRatings/HeroBannerRating";
 import HeroBannerTags from "../atoms/HeroBannerTags";
